+++ server/db/schemas/offer.js
@@ -1,6 +1,5 @@
 var db = require('../db');
 var Sequelize = require('sequelize');
-var Application = require('./application');
 
 var Offer = db.define('Offers', {
   application_id: Sequelize.INTEGER,
@@ -10,10 +9,4 @@ var Offer = db.define('Offers', {
   notes: Sequelize.TEXT
 });
 
-Offer.belongsTo(Application, {
-  foreignKey: application_id,
-  constraints: false
-});
-
-Offer.sync();
 module.exports = Offer;
