+++ server/db/schemas/phonescreen.js
@@ -1,6 +1,5 @@
 var db = require('../db');
 var Sequelize = require('sequelize');
-var Application = require('./application');
 
 var PhoneScreen = db.define('PhoneScreens', {
   application_id: Sequelize.INTEGER,
@@ -10,10 +9,4 @@ var PhoneScreen = db.define('PhoneScreens', {
   notes: Sequelize.TEXT
 });
 
-PhoneScreen.belongsTo(Application, {
-  foreignKey: application_id,
-  constraints: false
-});
-
-PhoneScreen.sync();
 module.exports = PhoneScreen;
